Allow a different position when on mobile
I like to use lhs on my desktop, but that doesn't seem to work/show when
on mobile. When on mobile, the modal or bhs position seems to work best,
so this allows a mobile position to be specified separately from the
normal desktop position.

Since the plug runs in a dedicated worker, it can't query for the
`pointer:fine` selector to determine if it's running on mobile like
silverbullet itself does. Because the worker is pretty limited, I've
opted to use what's available in the navigator object. The
userAgentData field should be available in most major browsers, but it's
not supported in Safari so I've opted to fallback on checking the user
agent string despite that approach not being approved by all.